Remove the blog post parsing to a new module
