+++ lib/elixir_newbie/blog/blog_parser.ex
@@ -1,0 +1,17 @@
+defmodule ElixirNewbie.Blog.BlogParser do
+  alias ElixirNewbie.Blog.Post
+
+  use NimblePublisher,
+    build: Post,
+    from: Application.app_dir(:elixir_newbie, "priv/posts/blog/**/*.md"),
+    as: :posts,
+    highlighters: [:makeup_elixir, :makeup_erlang]
+
+  @posts Enum.sort_by(@posts, & &1.date, {:desc, Date})
+
+  @tags @posts |> Enum.flat_map(& &1.tags) |> Enum.uniq() |> Enum.sort()
+
+  def posts_and_drafts, do: @posts
+
+  def all_tags, do: @tags
+end
